+++ zarrs_storage/src/storage_adapter.rs
@@ -2,6 +2,8 @@
 //!
 //! Storage adapters can be layered on stores.
 
+pub mod atomic_write;
+
 #[cfg(feature = "async")]
 pub mod async_to_sync;
 #[cfg(feature = "async")]